requirements.txt
nltk==3.7
numpy==1.23.3
Pillow==9.2.0
